Made some refixes to the code, to reduce the list sizing. Instead of only checking if the letters are invalid positions, which means we can only use yellow guesses (a yellow guess tells you a letter is present and that the current index is incorrect), I added green guesses in as well. This exponentially reduces time (26x26x...) as those letters are locked in place and have a known index.
I have also added in word validation at the end, which did not work with PyDictionary, so it has been changed to PyEnchant. This simply returns if a word is True or False (if it is a valid word in the given language - en_US for wordle as it is part of the NY Times, NY being a state within the USA).

This guesser is better than the last version (v2) as fake words don't count to final list shown to user.